\name{Data}
\alias{Q}
\alias{T2m.ncep}
\alias{Tt.prof}
\alias{Tz.prof}
\alias{atmOverturning}
\alias{levpres}
\alias{levpreseraint}
\alias{olr}
\alias{olr.t2m.corr}
\alias{olrtsi}
\alias{sky}
\alias{ta}
\alias{ta2}
\alias{wa}
\title{Data for demonstrations of the GHE and plotting related figures.}
\description{
  \code{Q} - heat flux.
  
\code{T2m.ncep} - 2m annual mean surface temperature from NCEP/NCAR
reanalysis. 
\code{Tt.prof} Vertical global mean temperature profile from
ERAINT \url{http://www.ecmwf.int}. \code{Tz.prof} is a subset of \code{Tt.prof} with only the troposphere.. 
\code{ta} \& \code{ta2} - global mean air temperature from ERAINT and
NCEP reanalyses

\code{wa} measure of anomalous vertical uppwelling \eqn{eta-z}{\eta_z(t)}.   

\code{atmOverturning} contains the measure of up and downselling -
vertical valocity. 

\code{levpres} - the pressure levels of the vertical levels in the
ERAINT reanalysis. \code{levpreseraint}

\code{olr} - Outgoing longwave radiation
(OLR). \url{http://www.cdc.noaa.gov/cdc/data.interp_OLR.html}. \code{olr.t2m.corr} - time series between OLR. \code{olrtsi} contains global mean OLR and total solar irradiance (TSI).

\code{sky} - a JPE figure used for the graphics background.

}
\usage{
data(Q)
data(T2m.ncep)
data(Tt.prof)
data(Tz.prof)
data(atmOverturning)
data(levpres)
data(levpreseraint)
data(olr)
data(olrtsi)
data(olr.t2m.corr)
data(sky)
data(ta)
data(ta2)
data(wa)
}

\value{Lists, data.frames, or vecors containing the relevant data. All
  the data have attributes providing meta-data information.}
\author{R.E. Benestad}
\keyword{data}

